Formatting
Former-commit-id: faab622ea5c68f614c102a39df0cb73472cf2e42

+++ src/com/fave100/server/domain/favelist/FaveList.java
@@ -284,13 +284,13 @@ public class FaveList extends DatastoreObject {
 	}
 
 	public static double calculateItemScore(final int position) {
-		return ((-1 * position) / 11 + (111 / 11));
+		return ((double)(-1 * position) / 11 + (111 / 11));
 	}
-	
-	public static List<String> getTrendingFaveLists() {		
+
+	public static List<String> getTrendingFaveLists() {
 		List<Hashtag> hashtags = ofy().load().type(Hashtag.class).order("zscore").limit(5).list();
 		List<String> trending = new ArrayList<>();
-		for(Hashtag hashtag : hashtags) {
+		for (Hashtag hashtag : hashtags) {
 			trending.add(hashtag.getName());
 		}
 		return trending;
